requirements.txt
# Install using `conda install -c conda-forge --file requirements.txt`
django==4.1
django-environ>=0.4.5
django-webpack-loader==1.4.1
psycopg2==2.9.3
whitenoise==6.4.0
gunicorn==20.1.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
